refactor(subway package): package by feature 형태로 변경
-feature(domain) 단위로 패키지를 변경함으로써 높은 응집성(또는 모듈성)의 장점을 살린다.
