--- src/main/java/nextstep/subway/ControllerExceptionHandler.java
@@ -1,21 +1,0 @@
-package nextstep.subway;
-
-import nextstep.subway.exception.DataNotFoundException;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class ControllerExceptionHandler {
-
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<Void> handleIllegalArgsException() {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(DataNotFoundException.class)
-    public ResponseEntity<Void> handleDataNotFoundException() {
-        return ResponseEntity.badRequest().build();
-    }
-}
