+++ src/main/java/nextstep/subway/common/exception/DataNotFoundException.java
@@ -1,0 +1,7 @@
+package nextstep.subway.common.exception;
+
+public class DataNotFoundException extends RuntimeException {
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+}
