+++ src/main/java/nextstep/subway/common/message/ExceptionMessage.java
@@ -1,0 +1,9 @@
+package nextstep.subway.common.message;
+
+public class ExceptionMessage {
+    public static final String REQUIRED = "필수값은 비어선 안됩니다!";
+    public static final String NOT_FOUND_LINE = "노선을 찾을 수 없습니다.";
+
+    private ExceptionMessage() {
+    }
+}
