--- src/main/java/nextstep/subway/exception/DataNotFoundException.java
@@ -1,7 +1,0 @@
-package nextstep.subway.exception;
-
-public class DataNotFoundException extends RuntimeException {
-    public DataNotFoundException(String message) {
-        super(message);
-    }
-}
