--- src/main/java/nextstep/subway/message/ExceptionMessage.java
@@ -1,9 +1,0 @@
-package nextstep.subway.message;
-
-public class ExceptionMessage {
-    public static final String REQUIRED = "필수값은 비어선 안됩니다!";
-    public static final String NOT_FOUND_LINE = "노선을 찾을 수 없습니다.";
-
-    private ExceptionMessage() {
-    }
-}
