--- src/main/java/nextstep/subway/ui/LineController.java
@@ -1,50 +1,0 @@
-package nextstep.subway.ui;
-
-import nextstep.subway.application.LineService;
-import nextstep.subway.dto.LineRequest;
-import nextstep.subway.dto.LineResponse;
-import nextstep.subway.dto.LineUpdateRequest;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-public class LineController {
-    private final LineService lineService;
-
-    public LineController(LineService lineService) {
-        this.lineService = lineService;
-    }
-
-    @PostMapping("/lines")
-    public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
-        LineResponse line = lineService.saveLine(lineRequest);
-        return ResponseEntity.created(URI.create("/lines/" + line.getId())).body(line);
-    }
-
-    @GetMapping(value = "/lines", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<LineResponse>> showLines() {
-        return ResponseEntity.ok().body(lineService.findAllLines());
-    }
-
-    @GetMapping(value = "/lines/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<LineResponse> showLine(@PathVariable(name = "id") Long id) {
-        return ResponseEntity.ok().body(lineService.findLine(id));
-    }
-
-    @PutMapping(value = "/lines/{id}")
-    public ResponseEntity<LineResponse> updateLine(@PathVariable(name = "id") Long id,
-                                                   @RequestBody LineUpdateRequest lineRequest) {
-        lineService.updateLine(id, lineRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping(value = "/lines/{id}")
-    public ResponseEntity<Void> deleteLine(@PathVariable(name = "id") Long id) {
-        lineService.deleteLine(id);
-        return ResponseEntity.noContent().build();
-    }
-}
