+++ src/test/java/nextstep/subway/line/LineAcceptanceMethods.java
@@ -3,8 +3,8 @@ package nextstep.subway.line;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceMethods;
-import nextstep.subway.dto.LineRequest;
-import nextstep.subway.dto.LineUpdateRequest;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineUpdateRequest;
 
 import static nextstep.subway.station.StationAcceptanceMethods.지하철역_생성;
